add text to sst rewriting support
